IBM Z: gen-vect-11/32: Set min-vect-loop-bound param back to default
In the Z backend we still set min-vect-loop-bound to 2 to work around
corner cases where awkward epilogue code gets generated in the
vectorizer.  This has a particular bad impact when vectorizing loops
with a low iteration count.  Due to this we do not vectorize the loop
in gen-vect-11/32 - what actually is a pity.

The patch sets min-vect-loop-bound back to the default value of 0 in
order to enable vectorization.

2019-11-05  Andreas Krebbel  <krebbel@linux.ibm.com>

	* gcc.dg/tree-ssa/gen-vect-11.c: Add --param min-vect-loop-bound=0
	for IBM Z.
	* gcc.dg/tree-ssa/gen-vect-11.c: Likewise.
	


git-svn-id: svn+ssh://gcc.gnu.org/svn/gcc/trunk@277831 138bc75d-0d04-0410-961f-82ee72b054a4