[Darwin, testsuite] Fix framework-1.c on later Darwin.
The test works by checking that a known framework path is accessible
when the '-F' option is given.  We need to find a framework path that
exists across a range of Darwin versions, and parseable by GCC.  This
adjusts the test to use a header path that exists and is parsable from
Darwin9 through Darwin19.

gcc/testsuite/ChangeLog:

2019-11-06  Iain Sandoe  <iain@sandoe.co.uk>

	* gcc.dg/framework-1.c: Adjust test header path.



git-svn-id: svn+ssh://gcc.gnu.org/svn/gcc/trunk@277892 138bc75d-0d04-0410-961f-82ee72b054a4